backend/requirements.txt
fastapi==0.110.1
uvicorn==0.27.1
langchain==1.0.5
langchain-openai==1.0.2
langgraph==1.0.3
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.31.0
pymongo==4.10.1
pandas==2.2.2
openai>=1.0.0